// src/components/Typography/types.ts
export type TSpacing =
  | "xxs"
  | "xs"
  | "s"
  | "sm"
  | "smm"
  | "m"
  | "l"
  | "xl"
  | "xxl"
  | "null";

export type TColor =
  | "primary"
  | "primary-70"
  | "primary-50"
  | "primary-20"
  | "accent"
  | "accent-50"
  | "white"
  | "white-60"
  | "white-50"
  | "active"
  | "error"
  | "error-70";